backend/requirements.txt
fastapi==0.109.2
uvicorn==0.30.1
python-dotenv==1.0.1
motor==3.3.2
bcrypt==4.1.2
user-agents==2.2.0
slowapi==0.1.1
pydantic==2.5.3
email-validator==2.0.0
PyYAML==6.0.1
pymongo==4.7.0
python-jose==3.3.0
passlib==1.7.4
dnspython==2.6.1

# OCR and Image Processing
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3

# OCR Options (choose one or more)
# pytesseract==0.3.10  # Uncomment for Tesseract OCR
# google-cloud-vision==3.4.4  # Uncomment for Google Vision API
# openai==1.3.3  # Uncomment for OpenAI Vision APIpython-dotenv
